Fix login/logout flow
Remove duplicate dataportal.views.globus function, specify some URL
paths in one place
