+++ hubmap/frontend/urls.py
@@ -1,8 +1,10 @@
+from django.conf import settings
 from django.urls import path
 
 from . import views
 
 urlpatterns = [
     path('', views.FrontendAppView.as_view(), name='index'),
-    path('login/', views.globus, name='globus'),
+    path(settings.LOGIN_REDIRECT_URL.lstrip('/'), views.set_user_cookie, name='set-user-cookie'),
+    path('logout/', views.logout, name='logout'),
 ]
